standardize all link to use same UI
